requirements.txt
streamlit==1.57.0
anthropic==0.97.0
openai==2.33.0
psycopg2-binary==2.9.12
pgvector==0.4.2
python-dotenv==1.2.2
sqlparse==0.5.5
